+++ cpu/x86/Makefile.x86
@@ -1,4 +1,4 @@
-CONTIKI_CPU_DIRS = . drivers
+CONTIKI_CPU_DIRS = . drivers/legacy_pc
 
 CONTIKI_SOURCEFILES += gdt.c helpers.S idt.c cpu.c rtc.c pit.c pic.c
